CanLookupAccount: create re-usable trait for managing an AccountLookup object
